Scripts to select and reduce data for Kr-like analysis.
Scripts in this commit read PMAP files, perform any needed selection
(e.g, selection on multiplicity), and write down a pandas DF with
column summary data (a KDST), ready for selection and histogramming.
The KDST contain the following variables.

run = run number
event = absolute event_number
timestamp = time stamp in ms
peak = peak number (counts the number of S2 in the event)
s1_energy = energy of S1 in PES
s2_energy = energy of S2 in PES (corresponding to peak number)
s1_epeak = energy of the S1 peak in PES (max of the energy for S1)
s2_epeak = energy of the S2 peak in PES (max of the energy for S2)
s1_width = width of S1 (in time units)
s2_width = width of S2 (in time units)
n_s1_pmt = number of S1 found in the PMTs that match the S1 found in the sum
n_sipm = number of SiPM found in the event
x = x coordinate (cartesian system)
y = y coordinate (cartesian system)
z = z coordinate (cartesian system)
r = radial coordinate (cylindrical system)
phi = azimuth coordinate (cylindrical system)

The structure of the Kryptonite script is essentially that of a city,
consuming PMAPS and producing a KDST. It is therefore a good candidate
to IC promotion.  However, the plan is to develop and debug it in
ICARO  before proposing a PR to IC
(one also needs to find whether this goes to main or to a specific branch).

The main driver, kryptonite.py loops over PMAP files and select events
of interest, which are written to file, via the writer interface
defined in kr_io.py. pmp_functions.py and hst_functions.py
provide a number of auxiliary functions.