Add notebook for XR filtering
This notebook takes a kDST and filters the data to produce
a XR-clean kDST in order to facilitate the Kr analysis.

In presence of external gamma sources, the XR production
is very significative and constitutes a source of
background for Kr data. This NB filters the data to
produce a rather pure dataset on which to perform
the regular analysis.